requirements.txt
cleantext
translators
openai-whisper
torch
torchvision
torchaudio
pytubefix
pydub